refactor\!: unify API naming and improve reducer clarity
BREAKING CHANGES:
- Rename `Reducer.withLock()` to `Reducer.lock()` for API consistency
- Rename `ReduceWithLock` to `LockmanDynamicConditionReducer` for clarity
- Consolidate implementation by removing ReduceWithLock.swift

The new naming better reflects the functionality and aligns with
Lockman's established naming patterns (LockmanDynamicConditionStrategy,
LockmanDynamicConditionInfo, etc.). The `lock()` method is now
consistently used across all APIs.
